feat: Modeullak Storage State Management
